Implemented maintenance mode
