+++ app/templates/503.html
@@ -1,0 +1,38 @@
+<!--
+SPDX-FileCopyrightText: 2025 PeARS Project, <community@pearsproject.org>, 
+
+SPDX-License-Identifier: AGPL-3.0-only
+-->
+
+{% extends "base/base.html" %}
+{% block body %}
+  <div class="row">
+    <div class="col-lg-2"></div>
+    <div class="col-lg-8 text-center mt-5">
+      {% if own_brand %}
+	{% if session.get('theme') == 'dark' %}
+          <a href="{{ url_for('search.index')}}"><img width=350px; src="{{url_for('serve_logos', path='logo-dark.png')}}"></a>
+	{% else %}
+          <a href="{{ url_for('search.index')}}"><img width=350px; src="{{url_for('serve_logos', path='logo.png')}}"></a>
+	{% endif %}
+      {% else %}
+	{% if session.get('theme') == 'dark' %}
+          <a href="{{ url_for('search.index')}}"><img width=350px; src="{{ url_for('static', filename='pears-logo-dark.png')}}"></a>
+	{% else %}
+          <a href="{{ url_for('search.index')}}"><img width=350px; src="{{ url_for('static', filename='pears-logo.png')}}"></a>
+	{% endif %}
+      {% endif %}
+	{% with messages = get_flashed_messages() %}
+        {% if messages %}
+            <div class="notification is-danger">
+		    <br><ion-icon name="hammer-outline"></ion-icon><ion-icon name="hammer-outline"></ion-icon><ion-icon name="hammer-outline"></ion-icon><br>
+		    <br><b>WORKING HARD</b><br>
+		    <b>{{ messages[0] }}</b><br>
+		    <br><ion-icon name="hammer-outline"></ion-icon><ion-icon name="hammer-outline"></ion-icon><ion-icon name="hammer-outline"></ion-icon>
+            </div>
+        {% endif %}
+        {% endwith %}
+    </div>
+  </div> 
+{% endblock %}
+
